move these classes to example package, because that's what they are...
