+++ nexus-app/src/main/java/com/github/nexus/api/example/SomeResource.java
@@ -1,15 +1,16 @@
-package com.github.nexus.api;
+package com.github.nexus.api.example;
 
 import com.github.api.nexus.quorum.v1.SomeObject;
-import java.util.Objects;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+
 import javax.validation.Valid;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.util.Objects;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  * Demonstration of an endpoint that uses a data object which is dynamically generated from an xsd.
